Reformat code: JetBrains Reformat code. Cast to u64/i64

+++ src/platform/apple/methods/error/from_raw.rs
@@ -1,7 +1,7 @@
 use crate::error::Error;
 
 impl Error {
-    pub fn from_raw(err: i32) -> Self {
+    pub fn from_raw(err: i64) -> Self {
         match err {
             1 => Self::PermissionDenied,
             2 => Self::NotFound,
