+++ src/platform/apple/methods/error/last.rs
@@ -5,7 +5,7 @@ impl Error {
         return Self::from_raw(Self::last_raw());
     }
 
-    pub fn last_raw() -> i32 {
-        return unsafe { *libc::__error() as i32 };
+    pub fn last_raw() -> i64 {
+        return unsafe { *libc::__error() as i64 };
     }
 }
