+++ src/platform/apple/methods/file/fs/file_metadata/size.rs
@@ -1,7 +1,7 @@
 use crate::file::FileMetadata;
 
 impl FileMetadata {
-    pub fn size(&self) -> i64 {
-        return self.metadata.st_size;
+    pub fn size(&self) -> u64 {
+        return self.metadata.st_size as u64;
     }
 }
