+++ src/platform/apple/prelude/write.rs
@@ -2,5 +2,5 @@ pub trait Write<B>: Sized
 where
     B: AsRef<str>,
 {
-    fn write(&mut self, buf: B) -> crate::error::Result<usize>;
+    fn write(&mut self, buf: B) -> crate::error::Result<u64>;
 }
