+++ src/platform/apple/types/file/seek_whence.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Copy, Clone)]
 pub enum SeekWhence {
-    StartPos(usize),
-    CurrentPos(isize),
-    EndPos(isize),
+    StartPos(u64),
+    CurrentPos(i64),
+    EndPos(i64),
 }
